backend/requirements.txt
Flask
Flask-Sockets
Flask-Cors
gevent
Flask-Bcrypt
PyJWT
python-dotenv